requirements.txt
django
djangorestframework
pandas
pymysql
langchain-community
python-dotenv
faiss-cpu

langchain-ollama

chromadb
langchain_experimental
tabulate
cryptography
